+++ src/raft.rs
@@ -1,5 +1,5 @@
 pub mod node;
 pub mod timer;
-mod storage;
+pub(crate) mod storage;
 pub mod rpc;
 pub mod proto;